Use shared libbase
Because the runtime apex already depends on libbase as shared object,
linkerconfig can use it as shared as well. This produces smaller
executable. On Cuttlefish, it reduces the size from 450632 to 328896.
It also affects the size of microdroid.img: 42893312 to 42668032.

Note that libbase is still statically linked on host since we want it to
have fewer external dependencies when running on host.

Bug: 262330207
Test: m linkerconfig
Test: MicrodroidTestApp
Change-Id: I3f2fd65626e270c3db30c892004ef068411dd10d